+++ lib/blip_ds.dart
@@ -24,6 +24,7 @@ export 'src/services/ds_auth.service.dart' show DSAuthService;
 export 'src/services/ds_bottom_sheet.service.dart' show DSBottomSheetService;
 export 'src/services/ds_dialog.service.dart' show DSDialogService;
 export 'src/services/ds_file.service.dart' show DSFileService;
+export 'src/services/ds_media_format.service.dart' show DSMediaFormatService;
 export 'src/services/ds_toast.service.dart' show DSToastService;
 export 'src/themes/colors/ds_colors.theme.dart' show DSColors;
 export 'src/themes/colors/ds_linear_gradient.theme.dart' show DSLinearGradient;
@@ -150,6 +151,8 @@ export 'src/widgets/utils/ds_cached_network_image_view.widget.dart'
     show DSCachedNetworkImageView;
 export 'src/widgets/utils/ds_card.widget.dart' show DSCard;
 export 'src/widgets/utils/ds_chip.widget.dart' show DSChip;
+export 'src/widgets/utils/ds_circular_progress.widget.dart'
+    show DSCircularProgress;
 export 'src/widgets/utils/ds_divider.widget.dart' show DSDivider;
 export 'src/widgets/utils/ds_expanded_image.widget.dart' show DSExpandedImage;
 export 'src/widgets/utils/ds_file_extension_icon.util.dart'