+++ lib/src/controllers/chat/ds_image_message_bubble.controller.dart
@@ -2,12 +2,12 @@ import 'dart:async';
 import 'dart:convert';
 import 'dart:io';
 
-import 'package:blip_ds/src/utils/ds_directory_formatter.util.dart';
 import 'package:crypto/crypto.dart';
 import 'package:get/get.dart';
 
 import '../../services/ds_auth.service.dart';
 import '../../services/ds_file.service.dart';
+import '../../utils/ds_directory_formatter.util.dart';
 
 class DSImageMessageBubbleController extends GetxController {
   final maximumProgress = RxInt(0);
@@ -39,24 +39,20 @@ class DSImageMessageBubbleController extends GetxController {
 
     final uri = Uri.parse(url);
 
-    final fullPath = await DSDirectoryFormatter.getPath(
+    final cachePath = await DSDirectoryFormatter.getCachePath(
       type: mediaType!,
-      fileName: md5.convert(utf8.encode(uri.path)).toString(),
+      filename: md5.convert(utf8.encode(uri.path)).toString(),
     );
 
-    if (await File(fullPath).exists()) {
-      localPath.value = fullPath;
+    if (File(cachePath).existsSync()) {
+      localPath.value = cachePath;
       return;
     }
 
-    final fileName = fullPath.split('/').last;
-    final path = fullPath.substring(0, fullPath.lastIndexOf('/'));
-
     try {
       final savedFilePath = await DSFileService.download(
-        url,
-        fileName,
-        path: path,
+        url: url,
+        path: cachePath,
         onReceiveProgress: _onReceiveProgress,
         httpHeaders: shouldAuthenticate ? DSAuthService.httpHeaders : null,
       );