+++ lib/src/widgets/chat/ds_file_message_bubble.widget.dart
@@ -42,7 +42,6 @@ class DSFileMessageBubble extends StatelessWidget {
   Widget build(BuildContext context) {
     return GestureDetector(
       onTap: () => controller.openFile(
-        filename: filename,
         url: url,
         httpHeaders: shouldAuthenticate ? DSAuthService.httpHeaders : null,
       ),